upd(hermes): token creation wrapped in API
